requirements.txt
fastapi
uvicorn
pydantic
pycryptodome
starlette
requests
websockets
ptyprocess; sys_platform != "win32"
pywinpty; sys_platform == "win32"
psutil
